+++ ext/duckdb/value.c
@@ -20,6 +20,7 @@ static VALUE duckdb_value_s__create_varchar(VALUE klass, VALUE str);
 static VALUE duckdb_value_s__create_blob(VALUE klass, VALUE str);
 static VALUE duckdb_value_s__create_hugeint(VALUE klass, VALUE lower, VALUE upper);
 static VALUE duckdb_value_s__create_uhugeint(VALUE klass, VALUE lower, VALUE upper);
+static VALUE duckdb_value_s__create_decimal(VALUE klass, VALUE lower, VALUE upper, VALUE width, VALUE scale);
 static VALUE duckdb_value_s_create_null(VALUE klass);
 
 static const rb_data_type_t value_data_type = {
@@ -126,6 +127,16 @@ static VALUE duckdb_value_s__create_uhugeint(VALUE klass, VALUE lower, VALUE upp
     uhugeint.lower = NUM2ULL(lower);
     uhugeint.upper = NUM2ULL(upper);
     duckdb_value value = duckdb_create_uhugeint(uhugeint);
+    return rbduckdb_value_new(value);
+}
+
+static VALUE duckdb_value_s__create_decimal(VALUE klass, VALUE lower, VALUE upper, VALUE width, VALUE scale) {
+    duckdb_decimal decimal;
+    decimal.value.lower = NUM2ULL(lower);
+    decimal.value.upper = NUM2LL(upper);
+    decimal.width = (uint8_t)NUM2UINT(width);
+    decimal.scale = (uint8_t)NUM2UINT(scale);
+    duckdb_value value = duckdb_create_decimal(decimal);
     return rbduckdb_value_new(value);
 }
 
@@ -274,6 +285,7 @@ void rbduckdb_init_duckdb_value(void) {
     rb_define_private_method(rb_singleton_class(cDuckDBValue), "_create_blob", duckdb_value_s__create_blob, 1);
     rb_define_private_method(rb_singleton_class(cDuckDBValue), "_create_hugeint", duckdb_value_s__create_hugeint, 2);
     rb_define_private_method(rb_singleton_class(cDuckDBValue), "_create_uhugeint", duckdb_value_s__create_uhugeint, 2);
+    rb_define_private_method(rb_singleton_class(cDuckDBValue), "_create_decimal", duckdb_value_s__create_decimal, 4);
     rb_define_singleton_method(cDuckDBValue, "create_null", duckdb_value_s_create_null, 0);
 }